refactor(loaders): rename 'session' to 'pretalx'
